House layout glitch fixed when opening/closing windows
Due to the way room information is now displayed (through a window which pops up when clicking on a room), the old way of displaying room data is no longer desired or necessary.

Before, any changes to the state of the house (rooms/windows specifically) would write an entirely new string in the layout room's box; now, we only need the room's number and name in the box. This commit prevents the layout room's text contents from changing when opening/closing windows/doors.