Heating/cooling model complete
Currently, it is now possible to change the temperature of a room manually (by setting the room temperature in the SHH tab), or by setting the outside temperature and opening a window while the simulation is running. The room with the open window will have its temperature increase/decrease by 0.1 degrees every second, until the temperature of the room is equal to the temperature outside.

Currently, some more features need to be implemented/tested, such as closing a window mid-temperature change, or pausing the simulator mid-temperature change.