Heating/cooling model glitches fixed
This code allows the simulator to be paused/stopped/resumed while the room temperature is adjusting to the target temperature.

If the simulator is running and a window is open, the room temperature will adjust towards the outside temperature. Pausing the simulator will pause the temperature timeline in TemperatureManager, and resuming the simulator will have the timeline play again. Pressing "Stop Simulator" will reset the room's temperature back to the initial 23.5 degrees celsius.